Abstract keyboard shorcuts for heading to paragraph transform and vice-versa.
