+++ packages/edit-site/src/components/keyboard-shortcuts/register.js
@@ -55,28 +55,6 @@ function KeyboardShortcutsRegister() {
 				},
 			],
 		} );
-
-		registerShortcut( {
-			name: 'core/edit-site/transform-heading-to-paragraph',
-			category: 'block-library',
-			description: __( 'Transform heading to paragraph.' ),
-			keyCombination: {
-				modifier: 'access',
-				character: `0`,
-			},
-		} );
-
-		[ 1, 2, 3, 4, 5, 6 ].forEach( ( level ) => {
-			registerShortcut( {
-				name: `core/edit-site/transform-paragraph-to-heading-${ level }`,
-				category: 'block-library',
-				description: __( 'Transform paragraph to heading.' ),
-				keyCombination: {
-					modifier: 'access',
-					character: `${ level }`,
-				},
-			} );
-		} );
 	}, [ registerShortcut ] );
 
 	return null;
